% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_fitness.R
\name{evaluate_fitness}
\alias{evaluate_fitness}
\title{Get Swarm Fitness}
\usage{
evaluate_fitness(X, fun)
}
\arguments{
\item{X}{A matrix of size \verb{N x d} representing the positions of the particles in the swarm.}

\item{fun}{A function that takes a numeric vector of length \code{d} as input and returns a numeric value (fitness).}
}
\value{
A numeric vector of length \code{N} representing the fitness values of the particles.
}
\description{
Get Swarm Fitness
}
\examples{
swarm <- initialize_swarm(10, 1, 2)
objective_function <- function(x) { sum(x^2) }
fitness <- evaluate_fitness(swarm$X, objective_function)
}
