% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_personal_bests.R
\name{initialize_personal_bests}
\alias{initialize_personal_bests}
\title{Initialize Matrices of Personal Bests}
\usage{
initialize_personal_bests(X, fitness)
}
\arguments{
\item{X}{A matrix of size \verb{N x d} representing the positions of the particles.}

\item{fitness}{A numeric vector of length \code{N} representing the current fitness values of the particles.}
}
\value{
A list containing two elements:
\item{P}{A matrix of size \verb{N x d} representing the personal best positions of the particles.}
\item{P_fitness}{A numeric vector of length \code{N} representing the personal best fitness values.}
}
\description{
Initialize Matrices of Personal Bests
}
\examples{
swarm <- initialize_swarm(30, c(-5, -5), c(5, 5))
fitness <- evaluate_fitness(swarm$X, function(x) sum(x^2))
personal_bests <- initialize_personal_bests(swarm$X, fitness)
}
\seealso{
\code{\link{initialize_swarm}}, \code{\link{evaluate_fitness}}
}
