% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_position.R
\name{update_position}
\alias{update_position}
\title{Update Particle Positions}
\usage{
update_position(X, V, xmin, xmax)
}
\arguments{
\item{X}{A matrix of size \verb{N x d} representing the current positions of the particles.}

\item{V}{A matrix of size \verb{N x d} representing the current velocities of the particles.}

\item{xmin}{Numeric vector of length \code{d}. The minimum bounds for each dimension.}

\item{xmax}{Numeric vector of length \code{d}. The maximum bounds for each dimension.}
}
\value{
A matrix of size \verb{N x d} representing the updated positions of the particles.
}
\description{
Updates the position of each particle in the swarm based on its current velocity.
The new positions are computed by adding the velocity to the current position.
If a position exceeds the boundaries defined by \code{xmin} and \code{xmax}, it is clipped.
}
\examples{
N <- 30
d <- 2
swarm <- initialize_swarm(N, c(-5, -5), c(5, 5))
V <- matrix(runif(N * d, -0.1, 0.1), nrow = N, ncol = d)
updated_positions <- update_position(swarm$X, V, c(-5, -5), c(5, 5))
}
\seealso{
\code{\link{initialize_swarm}}, \code{\link{update_velocity}}, \code{\link{evaluate_fitness}}
}
