% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_velocity.R
\name{update_velocity}
\alias{update_velocity}
\title{Update Swarm Velocities by FIPS}
\usage{
update_velocity(
  X,
  V,
  P,
  informants = lapply(1:nrow(X), function(i) sample(setdiff(1:nrow(X), i), size =
    floor(0.5 * nrow(X)) - 1)),
  w = 1/(2 * log(2)),
  c = 0.5 + log(2)
)
}
\arguments{
\item{X}{A matrix of size \verb{N x d} representing the current positions of the particles.}

\item{V}{A matrix of size \verb{N x d} representing the current velocities of the particles.}

\item{P}{A matrix of size \verb{N x d} representing the personal best positions of the particles.}

\item{informants}{A list of length \code{N}, where each element is a vector of indices representing the informants for each particle.}

\item{w}{Numeric. The inertia weight, which controls the influence of the current velocity.}

\item{c}{Numeric. The acceleration coefficient, which controls the influence of the personal best and social components.}
}
\value{
A matrix of size \verb{N x d} representing the updated velocities of the particles.
}
\description{
Update Swarm Velocities by FIPS
}
\examples{
N <- 30
swarm <- initialize_swarm(N, c(-5, -5), c(5, 5))
fitness <- evaluate_fitness(swarm$X, function(x) sum(x^2))
personal_bests <- initialize_personal_bests(swarm$X, fitness)
informants <- lapply(1:N, function(i) sample(setdiff(1:N, i), size = floor(0.5 * N) - 1))
updated_velocities <- update_velocity(swarm$X, swarm$V, personal_bests$P, informants)
head(updated_velocities )
}
\seealso{
\code{\link{initialize_swarm}}, \code{\link{initialize_personal_bests}}, \code{\link{update_personal_bests}}, \code{\link{update_position}}
}
